Move logging code to where it's used
